Tidy up recipe.js
Extracts the logic for displaying the various components of a recipe
into separate files
