+++ index.html
@@ -27,12 +27,12 @@
       <main role="main">
         <section class="index" itemscope itemtype="http://schema.org/Blog">
           <section class="posts clear">
-            <!-- element is filled by Javascript -->
+            <!-- element is filled by JS -->
           </section>
 
             <section class="posts clear bottom-panel" style="margin-bottom: 80px;">
               <div class="social clear">
-                <!-- element filled by Javascript -->
+                <!-- element filled by JS -->
               </div>
           </section>
         </section>
@@ -50,6 +50,6 @@
       </div>
     </footer>
 
-    <script type="module" src="../js/index.js"></script>
+    <script type="module" src="../js/index/index.js"></script>
   </body>
 </html>
