+++ js/index/display_thumbnail.js
