+++ js/index/generate_social_media_svg.js
