+++ js/index/index.js
@@ -1,5 +1,5 @@
-import recipe_data from "./recipe_data.js"
-import socialMediaSvgData from "./social_media_svg_data.js"
+import recipe_data from "../recipe_data.js"
+import socialMediaSvgData from "../social_media_svg_data.js"
 
 import { displayThumbnail } from "./display_thumbnail.js"
 import { generateSocialMediaSvg } from "./generate_social_media_svg.js"
