+++ recipe.html
@@ -13,7 +13,7 @@
     <link href="https://fonts.googleapis.com/css?family=Cardo&display=swap" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Lora|Raleway:500" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css?family=Open+Sans" rel="stylesheet">
-    <script type="module" src="../js/recipe.js"></script>
+    <script type="module" src="../js/recipe/recipe.js"></script>
   </head>
 
   <body>
@@ -60,17 +60,11 @@
         <div class="post-content" itemprop="text">
           <div class="recipe-image"><!-- element is filled by JS --></div>
           <div class="introduction-paragraphs">
-            <div class="first-paragraph">
-              <p>Brussel sprouts have a reputation for being hated, and I think the reason for this is that many people simply have not had them prepared properly, or with enough love and finesse, to fully realise their potential. When simply boiled, they're not great. Because they're so dense, by the time the heart is cooked the vibrant green outer leaves have turned brown and unappetising. If you remove the outer leaves and cook them separately, the hearts can be slowly braised, creating a rich sauce that will infuse flavour into the hard centres.</p>
-            </div>
+            <div class="first-paragraph"><p><!-- element is filled by JS --></p></div>
           </div>
-
           <hr />
-
           <ul><!-- element is filled by JS --></ul>
-
           <hr />
-
           <ol><!-- element is filled by JS --></ol>
         </div>
       </article>
@@ -88,6 +82,6 @@
       </div>
     </footer>
 
-    <script type="module" src="../js/recipe.js"></script>
+    <script type="module" src="../js/recipe/recipe.js"></script>
   </body>
 </html>
